+++ CHANGELOG.md
@@ -15,6 +15,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Fix: Change transition on `k-RadioButton`.
 - Fix: Change `k-LinkBox` display before `m` version.
 - Fix: Change `k-SingleColumn` width according to the media.
+- Feature: Add new `Title` React component.
+- Feature: Add new `Paragraph` React component.
+- Feature: Add new `HeartIcon` React component.
+- Feature: Add new `Button` React component.
 - Feature: Add new `HeaderLogo` React component.
 - Fix: Change default port in development from 3000 to 3003. You should now
   see the style guide on http://localhost:3003 instead.